Safari-609.1.11.1/LayoutTests/css3/calc/color-hsl-expected.txt

Tests that CSS3 calc() can be used with the hsl() and hsla() functions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.getElementById("simple"), null).color is "rgb(32, 223, 32)"
PASS getComputedStyle(document.getElementById("alpha"), null).color is "rgba(32, 223, 32, 0.7)"
PASS successfullyParsed is true

TEST COMPLETE
